add socket read/write to runtime
- runtime now listens for messages from server
- when it gets message type 0, it walks the tu's and writes their current line
  execution counts to the socket
- also add support in the testsuite for this
  - runtime will exit if it can't connect to a viewer
- also add new test that gets execution counts and makes sure the lines make
  sense

+++ include/scv_global.h
@@ -1,7 +1,9 @@
+#include <stdint.h>
+
 struct scv_node {
-	unsigned int *lines_ptr;
-	unsigned int size;
+	/* long long in C99 is also guaranteed to be 64 bits */
+	uint64_t *lines_ptr;
+	uint64_t size;
 	const char *file_name;
 	struct scv_node *next;
-	/* unsigned int not_end; */
 };
